fix for get age and bioage commands, fix for bind PredictRequest

+++ src/Services/Agents.API/Agents.API.Service/Services/AgentInitSettingsProvider.cs
@@ -57,8 +57,8 @@ namespace Agents.API.Service.Services
                             },
                 Properties = new Dictionary<string, IProperty>
                             {
-                                { "CurrentAge", new AgentProperty("CurrentAge", typeof(long)) },
-                                { "CurrentBioAge", new AgentProperty("CurrentBioAge", typeof(long)) },
+                                { "CurrentAge", new AgentProperty("CurrentAge", typeof(int)) },
+                                { "CurrentBioAge", new AgentProperty("CurrentBioAge", typeof(int)) },
                                 { "CurrentAgeRang", new AgentProperty("CurrentAgeRang", typeof(AgentBioAgeStates)) }
                             },
                 StateDiagram = new StateDiagram(states, async x =>
@@ -67,7 +67,7 @@ namespace Agents.API.Service.Services
                     AgentBioAgeStates rang = Enum.Parse<AgentBioAgeStates>(x.Properties["CurrentAgeRang"].Value.ToString());
                     IAgentState state = states[rang.GetDisplayAttributeValue()];
                     state.Timestamp = x.Timestamp;
-                    state.NumericCharacteristic = (long)x.Properties["CurrentBioAge"].Value - (long)x.Properties["CurrentAge"].Value;
+                    state.NumericCharacteristic = (int)x.Properties["CurrentBioAge"].Value - (int)x.Properties["CurrentAge"].Value;
                     return state;                    
                 })
             };
